chore(vscode-extension): merge utils and helpers files (#47)
